+++ src/groovy/net/hedtech/restfulapi/config/RepresentationDelegate.groovy
@@ -1,5 +1,5 @@
 /* ****************************************************************************
- * Copyright 2013 Ellucian Company L.P. and its affiliates.
+ * Copyright 2013-2015 Ellucian Company L.P. and its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ class RepresentationDelegate {
     String marshallerFramework
     String contentType
     def jsonArrayPrefix
+    def patchSupport
     def marshallers = []
     def extractor